(trunk) Replaced source:/gplates/trunk using source:/gplates/branches/raster-in-map-view-2011-dec-5@13070 incorporating log:/gplates/branches/raster-in-map-view-2011-dec-5@12723:13070
This is "merging one branch to another without previous sync". The merge was performed using Tortoise SVN's Merge tool, from[[br]]
source:/gplates/branches/raster-in-map-view-2011-dec-5@12722 to[[br]]
source:/gplates/branches/raster-in-map-view-2011-dec-5@13070, [[br]]
from inside the root directory of a working copy of the trunk.

git-svn-id: http://svn.gplates.org/projects/gplates/trunk@13071 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c

+++ src/opengl/GLMultiResolutionRasterInterface.cc
@@ -51,7 +51,7 @@ GPlatesOpenGL::GLMultiResolutionRasterInterface::render(
 		float level_of_detail_bias)
 {
 	// Get the level-of-detail based on the size of viewport pixels projected onto the globe.
-	const int level_of_detail =
+	const float level_of_detail =
 			clamp_level_of_detail(
 					get_level_of_detail(
 							renderer.gl_get_matrix(GL_MODELVIEW),
